requirements.txt
pybit
pandas
numpy
python-telegram-bot[job-queue]